+++ docs/src/composables/dark.js
@@ -1,4 +1,0 @@
-// import { useDark, useToggle } from '@vueuse/core'
-
-export const isDark = VueUse.useDark()
-export const toggleDark = VueUse.useToggle(isDark)