+++ docs/src/composables/dark.mjs
@@ -1,0 +1,9 @@
+// import { useDark, useToggle } from '@vueuse/core'
+
+// const { useDark, useToggle} = 
+
+const isDark = VueUse.useDark()
+const toggleDark = VueUse.useToggle(isDark)
+
+// module.exports = {isDark,toggleDark}
+export { isDark, toggleDark }